packages/market-pulse/skills/market-pulse-brain/requirements.txt
PyYAML==6.0.2
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0
googletrans==4.0.0-rc1
